ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
joblib==1.3.2
pydantic==2.4.2
python-dateutil==2.8.2
kagglehub==0.1.5
plotly==5.17.0
matplotlib==3.8.0
seaborn==0.13.0
xgboost==2.0.0
catboost==1.2.2
lightgbm==4.1.0
